fix: move enum to shared
