+++ shared/models/index.ts
@@ -1,2 +1,3 @@
 export * from './db'
 export * from './user'
+export * from './notifications'
