+++ shared/models/notifications.ts
@@ -1,0 +1,6 @@
+export enum EmailNotificationFrequency {
+  NEVER = 'never',
+  DAILY = 'daily',
+  WEEKLY = 'weekly',
+  MONTHLY = 'monthly',
+}
